Added Race component and created 'Dwarf' race
Now the characters can instantiate a race in the PlayerCharacter component. So far, the race properties doesn't have much information. Still need to orgnize the race information. The Race class serves as the template for each of the races that needs to be implemented. It's defined as an abstract class so that different races can have different properties, but still follow the same basic implementation.