Added Integration between Class and PlayerCharacter
This commit servers as a integration between the PlayerCharacter and the Classes that are available in the game. I made the PlayerCharacter component store the classes as an array since it's possible to multi-class in the game. This should allow multiple classes to be stored in the PlayerCharacter component. But I'll need to limit the number of classes possible since there are some restrictions on which classes can be multi-classes into. I've also provided the basic functionality of the classes by creating an abstract class that will serve as the template for all the other CharacterClasses. I've also named the Character Classes as CharacterClass since I can't really call them Class (It's a programming thing. I can't use up already existing keywords). So far, the project still looks daunting, but well, slowly slowly.