Added Ability Score and Modifier class
Another essential part of the game, ability scores. I decided to create a class to manage all the responsibilities the ability scores are supposed to do. The ability scores themselves are going to be added to the playerCharacters class, allowing the characters themselves to reference the ability scores and their modifiers directly. So far I've only done the scores and the modifiers. Still need a way to manage the skills that are associated with it, the saving throw and any attacks that are tied to the scores (Spellcasting is goign to be a nightmare).

Also, I'm really not sure if creating a whole bunch of getters for the ability scores is a good design. But for now, I'll keep it like this and refactor it if it get's too annoying.