Added Abilities Integration to PlayerCharacter Component
Now the PlayerCharacters have abilities that further define the character. The ability socres themselves need a little bit more tweaking and I'm planning on adding more test scenarios in the PlayerCharacterAbilitiesTest integration test. Need to cover the modifiers functions. But I'm gonna change how the ability modifiers are going to be called. So for now, here's the integration between the two.