Added Levels feature
This commit adds the levels tracking feature. The feature itself is very very bare-bones, since it doesn't account for the different levels each class can have. For now, it's just a basic interface to keep track of what proficiency bonus and the level that your character is at based purely on the experience points that the character gets.