Moved Levels functionality to CharacterClass components
In D&D, the levels of a character can be tied to the class that the character has. It's also possible to multi-class, meaning picking two or more classes for their character.
I moved the levels functionality so that each class can track it's own Level and retrieve the proficiency bonus based on the level of that particular class. There's a few more details to iron out, like how will the experience points be allocated, how do we track different experience points in three different classes? etc.