Added Size functionality to Race
The selected race in the player handbook dictates the size of your character. I forsee it will play an important part when determinine combat and certain checks. Therefore it makes sense to extract it out into it's own class like Levels. So far, I'm still slowly implementing the various features available in each of the races. So far so good.