Added Speed and Proficiencies to Player Character
The speed class manages the speed of the player at all times. I've only managed to create a function that adds a modifier to the speed, but it should be enough to manipulate the speed value. I've also added proficiencies to the player character. So far, it's a simple class that is storing all the proficiencies the player has. Once again, a simple interface. There's not much checking if the proficiency even exists. We can implement that in the future. For now, it should be good enough to expand the rest of the program with.