Added Spell components
I feel one of the more useful features of this program is to have a spell book. This spell book should be able to allow the DM to reference spells quickly by inputting the spell name. So I'm fast-tracking this feature. I've downloaded a json file that contains all the spells in D&D 5e. So far I've managed to create a class to maintain the details of the spell. Next I'm gonna focus on creating a spellbook, a component that will load all the spells into the Spell objects and maintain the list while the program runs.