Added Spellbook component
After creating the spells component, I've now added the spellbook component. It's suppose to read all the spells in the list and load them into an array that can be searched by name. There's a bit more I'd like to do with the spellbook, like search spells for certain classes, organize the components of the spell into enumerated values and search the descriptions for keywords. But for now, it should be good enough to start building the CLI Interface component from. Since the spellbook should be functional in it's base form, the CLI Interface can now be designed to revolve around it. Gotta remember to make the CLI Interface scalable to the other components as well.