Changed the name 'Trait' to 'Feature'. Completed base for Features.
So now, the player character can now have 'Features' that come with their class and race. For now, I've figured out that we can keep the features in a feature list component that manages the features that a Player character can have. I've also added more into the setRace function and added a few cases that may arise. The dependencies of the abilities to the race is an inherent problem, but can be mitigated in the future.