Added Non Player Characters
I thought it might be a little more useful to create the Non Player Character component. I feel that it'll help to standardize more items in the D&D ruleset, allowing it to be easily transferred to code logic. So far, I've setup the first couple of items from the Monster manuals statistics block. We still need Armor Class and Hit Points and many other items. For now, let's start off with the basics.