Added console support
Added a rudimentary support for displaying all the spells in the
console. Still figuring out the kinks in using Symphony in all this, but
it's one step closer to what the tool is actually supposed to be about.
We still have a lot to think about, like how we're going to structure
the vast amount of commands required to run this entire ecosystem. For
now, we can just use the basic Command pattern accompanied by Symphony.
Work smarter, not harder, ammirite?

It's also worth noting that there's not a convienient way of creating
spells to test. We need extend the Spell component a bit. Make it a
little less dumb.